refactor(transactions): split page into modular components and extract helpers
