+++ lib/transactions/formatMoney.ts
@@ -1,0 +1,2 @@
+export const formatMoney = (value: number) =>
+  `${value < 0 ? '-' : ''}$${Math.abs(value).toFixed(2)}`
